Address critical27 and darion's comments

+++ src/jni/java/src/main/java/com/vesoft/nebula/NebulaCodec.java
@@ -91,7 +91,7 @@ public class NebulaCodec {
 
     public static native byte[] encode(Object[] values);
 
-    public static native List<byte[]> decode(byte[] encoded, Pair[] fields, int version);
+    public static native List<byte[]> decode(byte[] encoded, Pair[] fields, long version);
 
     private boolean checkKey(String key) {
         return Objects.isNull(key) || key.length() == 0;
